+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Add MultiMC instance importing [#426]
 - Allow setting which account to use in Instance settings
 - Add setting to hide pack name and version from Instance card
+- Add support for modpacks.ch (FTB) packs [#392]
 
 ### Fixes
 - Fix exported instances including disabled mods [#420]